feat!: Version set unions as solvable requirements (#56)
This PR introduces the concept of version set unions and allows them to be used to specify solvable requirements that can be fulfilled by more than one version set (belonging to more than one package).

The change is facilitated by the introduction of a Requirement type, which is used instead of VersionSetIds to specify a required dependency of a solvable. A requirement can either be comprised of a single version set (a VersionSetId), or multiple version sets (a VersionSetUnionId). This allows existing code bases to continue using a single version set (VersionSetId) as a requirement specification without ever having to touch version set unions (VersionSetUnionId).